chaged type of visited and pruned states from uint32_t to uint64_t with logs format string correction

+++ src/model/game/player/Player.cpp
@@ -33,12 +33,12 @@ std::string Player::getInfo() const
     return "Color: " + colorToStr(color) + ", name: " + name;
 }
 
-uint32_t Player::getVisitedStates() const
+uint64_t Player::getVisitedStates() const
 {
     return visitedStates;
 }
 
-uint32_t Player::getPrunedStates() const
+uint64_t Player::getPrunedStates() const
 {
     return prunedStates;
 }
